requirements.txt
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
Faker==33.1.0
fonttools==4.55.2
kiwisolver==1.4.7
matplotlib==3.9.3
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pymongo==4.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
typing_extensions==4.12.2
tzdata==2024.2
